+++ Classes/SBJson5StreamWriter.m
@@ -1,5 +1,5 @@
 /*
- Copyright (c) 2010, Stig Brautaset.
+ Copyright (c) 2010-2020, Stig Brautaset.
  All rights reserved.
 
  Redistribution and use in source and binary forms, with or without
@@ -35,7 +35,120 @@
 #endif
 
 #import "SBJson5StreamWriter.h"
-#import "SBJson5StreamWriterState.h"
+
+@class SBJson5StreamWriterState;
+
+@interface SBJson5StreamWriter ()
+@property (nonatomic, strong) SBJson5StreamWriterState *stateObjectStart,
+  *stateObjectKey,
+  *stateObjectValue,
+  *stateArrayStart,
+  *stateArrayValue,
+  *state;
+@property (nonatomic, readonly, strong) NSMutableArray *stateStack;
+@end
+
+#pragma mark -
+
+@interface SBJson5StreamWriterState : NSObject
+- (BOOL)isInvalidState:(SBJson5StreamWriter *)writer;
+- (void)appendSeparator:(SBJson5StreamWriter *)writer;
+- (BOOL)expectingKey:(SBJson5StreamWriter *)writer;
+- (void)transitionState:(SBJson5StreamWriter *)writer;
+- (void)appendWhitespace:(SBJson5StreamWriter *)writer;
+@end
+
+@interface SBJson5StreamWriterStateObjectStart : SBJson5StreamWriterState
+@end
+
+@interface SBJson5StreamWriterStateObjectKey : SBJson5StreamWriterStateObjectStart
+@end
+
+@interface SBJson5StreamWriterStateObjectValue : SBJson5StreamWriterState
+@end
+
+@interface SBJson5StreamWriterStateArrayStart : SBJson5StreamWriterState
+@end
+
+@interface SBJson5StreamWriterStateArrayValue : SBJson5StreamWriterState
+@end
+
+@interface SBJson5StreamWriterStateStart : SBJson5StreamWriterState
+@end
+
+@interface SBJson5StreamWriterStateComplete : SBJson5StreamWriterState
+@end
+
+#pragma mark -
+
+@implementation SBJson5StreamWriterState
+- (BOOL)isInvalidState:(SBJson5StreamWriter *)writer { return NO; }
+- (void)appendSeparator:(SBJson5StreamWriter *)writer {}
+- (BOOL)expectingKey:(SBJson5StreamWriter *)writer { return NO; }
+- (void)transitionState:(SBJson5StreamWriter *)writer {}
+- (void)appendWhitespace:(SBJson5StreamWriter *)writer {
+  [writer appendBytes:"\n" length:1];
+  for (NSUInteger i = 0; i < writer.stateStack.count; i++)
+    [writer appendBytes:"  " length:2];
+}
+@end
+
+@implementation SBJson5StreamWriterStateObjectStart
+- (void)transitionState:(SBJson5StreamWriter *)writer {
+    writer.state = writer.stateObjectValue;
+}
+- (BOOL)expectingKey:(SBJson5StreamWriter *)writer {
+  writer.error = @"JSON object key must be string";
+  return YES;
+}
+@end
+
+@implementation SBJson5StreamWriterStateObjectKey
+- (void)appendSeparator:(SBJson5StreamWriter *)writer {
+  [writer appendBytes:"," length:1];
+}
+@end
+
+@implementation SBJson5StreamWriterStateObjectValue
+- (void)appendSeparator:(SBJson5StreamWriter *)writer {
+  [writer appendBytes:":" length:1];
+}
+- (void)transitionState:(SBJson5StreamWriter *)writer {
+    writer.state = writer.stateObjectKey;
+}
+- (void)appendWhitespace:(SBJson5StreamWriter *)writer {
+  [writer appendBytes:" " length:1];
+}
+@end
+
+@implementation SBJson5StreamWriterStateArrayStart
+- (void)transitionState:(SBJson5StreamWriter *)writer {
+    writer.state = writer.stateArrayValue;
+}
+@end
+
+@implementation SBJson5StreamWriterStateArrayValue
+- (void)appendSeparator:(SBJson5StreamWriter *)writer {
+  [writer appendBytes:"," length:1];
+}
+@end
+
+@implementation SBJson5StreamWriterStateStart
+- (void)transitionState:(SBJson5StreamWriter *)writer {
+  writer.state = [[SBJson5StreamWriterStateComplete alloc] init];
+}
+- (void)appendSeparator:(SBJson5StreamWriter *)writer {
+}
+@end
+
+@implementation SBJson5StreamWriterStateComplete
+- (BOOL)isInvalidState:(SBJson5StreamWriter *)writer {
+  writer.error = @"Stream is closed";
+  return YES;
+}
+@end
+
+#pragma mark -
 
 @implementation SBJson5StreamWriter {
     BOOL _sortKeys, _humanReadable;
@@ -56,7 +169,7 @@
            humanReadable:(BOOL)humanReadable
                 sortKeys:(BOOL)sortKeys
       sortKeysComparator:(NSComparator)sortKeysComparator {
-    return [[self alloc] initWithDelegate:delegate
+  return [[self alloc] initWithDelegate:delegate
                                  maxDepth:maxDepth
                             humanReadable:humanReadable
                                  sortKeys:sortKeys
@@ -75,13 +188,19 @@
         kTrue = [NSNumber numberWithBool:YES];
         kFalse = [NSNumber numberWithBool:NO];
 
+        _stateObjectStart = [[SBJson5StreamWriterStateObjectStart alloc] init];
+        _stateObjectKey = [[SBJson5StreamWriterStateObjectKey alloc] init];
+        _stateObjectValue = [[SBJson5StreamWriterStateObjectValue alloc] init];
+        _stateArrayStart = [[SBJson5StreamWriterStateArrayStart alloc] init];
+        _stateArrayValue = [[SBJson5StreamWriterStateArrayValue alloc] init];
+        _state = [[SBJson5StreamWriterStateStart alloc] init];
+
         _delegate = delegate;
 		_maxDepth = maxDepth;
         _sortKeys = sortKeys;
         _humanReadable = humanReadable;
         _sortKeysComparator = sortKeysComparator;
         _stateStack = [[NSMutableArray alloc] initWithCapacity:maxDepth];
-        _state = [SBJson5StreamWriterStateStart sharedInstance];
         cache = [[NSMutableDictionary alloc] initWithCapacity:32];
     }
 	return self;
@@ -140,7 +259,7 @@
 	if (_humanReadable && _stateStack.count) [_state appendWhitespace:self];
 
     [_stateStack addObject:_state];
-    self.state = [SBJson5StreamWriterStateObjectStart sharedInstance];
+    self.state = self.stateObjectStart;
 
 	if (_maxDepth && _stateStack.count > _maxDepth) {
 		self.error = @"Nested too deep";
@@ -173,7 +292,7 @@
 	if (_humanReadable && _stateStack.count) [_state appendWhitespace:self];
 
     [_stateStack addObject:_state];
-	self.state = [SBJson5StreamWriterStateArrayStart sharedInstance];
+    self.state = self.stateArrayStart;
 
 	if (_maxDepth && _stateStack.count > _maxDepth) {
 		self.error = @"Nested too deep";